pull tree traversal methods into a trait
